New composable file for sso
